modified input field,added to export
