+++ lib/ui/views/authInputTestView/authInputTestView.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
 import 'package:zc_desktop_flutter/ui/views/authInputTestView/inputTestViewModel.dart';
-import 'package:zcdesk_ui/src/widgets/auth_input_field.dart';
-import 'package:zcdesk_ui/src/shared/ui_helpers.dart';
+import 'package:zcdesk_ui/zcdesk_ui.dart';
 
 class AuthInputTestView extends StatelessWidget {
   const AuthInputTestView({Key? key}) : super(key: key);
@@ -23,7 +22,7 @@ class AuthInputTestView extends StatelessWidget {
                     AuthInputField(
                       label: 'Email',
                       controller: model.emailController,
-                      hintText: 'Markhng@gmail.com',
+                      hintPlaceHolder: 'Markhng@gmail.com',
                       onSaved: (val) => model.onSaved(val),
                     ),
 
@@ -31,14 +30,11 @@ class AuthInputTestView extends StatelessWidget {
 
                     AuthInputField(
                       label: 'Password',
-                      hintText: 'Password',
+                      hintPlaceHolder: 'Password',
                       controller: model.passWordController,
                       onSaved: (val) => model.onSaved(val),
-                      suffix: TextButton(
-                        child: Icon(Icons.visibility_off_outlined,size: 14,),
-                        onPressed: () {
-
-                        },
+                      trailing: Icon(
+                          Icons.visibility_off_outlined,size: 16
                       ),
                       ),
