+++ lib/ui/views/authInputTestView/inputTestViewModel.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/cupertino.dart';
 import 'package:stacked/stacked.dart';
+import 'package:flutter/material.dart';
 
 class InputTestViewModel extends BaseViewModel{
   TextEditingController _emailController = TextEditingController();
