+++ packages/zcdesk_ui/lib/zcdesk_ui.dart
@@ -2,9 +2,11 @@ library zcdesk_ui;
 
 // widgets exports
 export 'src/widgets/zcdesk_text.dart';
+export 'src/widgets/zcdesk_input_field.dart';
 
 // colors export
 export 'src/shared/app_colors.dart';
 
 // shared exports
 export 'src/shared/ui_helpers.dart';
+
